refactor(checks): lift the three-state check out of the ladder
`pf.onboard.ladder` defined `Check(name, status, evidence)` and its three
constructors, and they are the right vocabulary for anything that reports a
derived verdict — not just an onboarding stage. The next caller needs them
without needing `ladder`, which imports `dialect`, `survey` and `audit` to do
its own job.

Two options were both worse. Importing from `ladder` drags three modules in to
reuse a twelve-line dataclass. Defining a second copy makes three, because
`pf.loops.audit.Check` already exists and is a different thing — a weighted
pass/fail for a readiness *score*, not a gate condition with evidence.

So it moves to `pf.checks` and `ladder` re-exports it. No behaviour change:
`Status`, `Check`, `passed`, `failed` and `unexercised` are importable from
exactly where they were, which is why the re-export carries a `noqa` rather than
being tidied away — removing a public name is a breaking change even when
nothing in this repository imports it yet.

The docstring keeps the reasoning that was scattered in `ladder`'s: why three
states rather than two, and why only `fail` closes a gate.
