+++ .gitmodules
@@ -83,3 +83,8 @@
 	path = vendor/public-sector-ai-playbook
 	url = https://github.com/PackMaaan/public-sector-ai-playbook
 	branch = main
+	shallow = true
+[submodule "vendor/ai-governance-framework"]
+	path = vendor/ai-governance-framework
+	url = https://github.com/Atomz-org/ai-governance-framework
+	branch = main
