+++ platform/src/pf/onboard/ladder.py
@@ -49,15 +49,13 @@ from collections.abc import Callable
 from dataclasses import dataclass, field
 from functools import cached_property
 from pathlib import Path
-from typing import Literal
 
 import yaml
 
+from pf.checks import Check, Status, failed, passed, unexercised  # noqa: F401 — Status re-exported
 from pf.onboard.audit import Risk
 from pf.onboard.dialect import Report, analyse
 from pf.onboard.survey import RESERVED_MACROS
-
-Status = Literal["pass", "fail", "unexercised"]
 
 #: The layers this platform reasons about. `semantic` is not a model directory in
 #: the same sense as the others — it holds MetricFlow YAML — but it lives under
@@ -78,34 +76,11 @@ _REF = re.compile(r"\{\{\s*(?:ref|source)\s*\(([^)]*)\)", re.IGNORECASE)
 
 
 # ------------------------------------------------------------------ checks --
-@dataclass(frozen=True)
-class Check:
-    """One gate condition, with the evidence for its verdict.
-
-    `evidence` is not a restatement of `name`. A check that says only "failed"
-    forces whoever reads it to reproduce the check by hand, which is how a gate
-    stops being read at all.
-    """
-
-    name: str
-    status: Status
-    evidence: str
-
-    @property
-    def ok(self) -> bool:
-        return self.status != "fail"
-
-
-def passed(name: str, evidence: str) -> Check:
-    return Check(name, "pass", evidence)
-
-
-def failed(name: str, evidence: str) -> Check:
-    return Check(name, "fail", evidence)
-
-
-def unexercised(name: str, evidence: str) -> Check:
-    return Check(name, "unexercised", evidence)
+# `Check`, `Status`, `passed`, `failed` and `unexercised` now live in
+# `pf.checks` — imported above and re-exported here, because this module defined
+# them first and `from pf.onboard.ladder import Check` is what the callers write.
+# `pf.air` needed the same vocabulary without this module's imports; the move is
+# the whole change.
 
 
 @dataclass
